Added shell script to handle collection session ending (via PUT request) * Shell script looks at the repository, creates a new folder labeled 'data' and adds a JSON file representing the data collected * Adjusted views.py to change the interaction with the PUT HTTP request

+++ collect/views.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 from django.shortcuts import render
 
 # Create your views here.
+import subprocess
 from django.http import HttpResponse, JsonResponse
 from rest_framework.decorators import api_view
 from .models import Biometrics
@@ -22,12 +23,15 @@ def get_hb_data (request):
     
     # Updates the list of instances with a new value
     elif request.method == 'PUT':
-        hb_data = Biometrics.objects.all().delete()
-        serializer = BiometricSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        process = subprocess.Popen(['./makeData.sh'], stdout=subprocess.PIPE)
+        message = process.communicate()[0].decode()
+        message = message.split("\n")[-2]
+
+        if process.returncode == 0:
+            hb_data = Biometrics.objects.all().delete()
+            return Response({"local_save": "successful", "type" : "JSON", "file_name" : message}, status=200)
+        
+        return Response({"local_save": "undefined", "type" : "JSON", "file_name" : message}, status=400)
     
     # Appends to the existing list of instances
     elif request.method == 'POST':
